LPS-69533 Move to counter package + SF
